Improved performance of STL loaders.
Both binary and ASCII STL loaders were O(n^2) in the number of vertices in the
input.  This change replaces both algorithms with O(n) equivalents, and
eliminates the need for a separate 'building faces' pass.  For a 10,238-polygon
model, this reduces load time by 93% on my machine (from 24 seconds to 1.5).