Removed unused normals array from face data.
It appears that the loader and viewer used to trust the normals
from the STL file, but at some point stopped doing this, in favor
of recomputing them on the fly.  Until this commit, they were
still passing around vestigial normal arrays, confusing people
like me who were trying to understand the code.