Handle files without space after 'solid'.
For ASCII STL files that have no names, most tools still put a
single space after 'solid' and 'endsolid'.  We've seen a set of
files in the wild, however, where the spaces are missing:

http://www.thingiverse.com/thing:4398

This change tweaks the regexps to tolerate such files (and also
regularizes the patterns for solid and endsolid, which I had
accidentally caused to diverge last week).